smash-config package refactor - tidy index and move utils into folder
