+++ packages/smash-config/src/utils/getSmashConfig.ts
@@ -1,4 +1,5 @@
-import type { SCSSAliases, SmashConfig } from "./index.js";
+import type { SCSSAliases, SmashConfig } from "../types.js";
+
 import { normalize, resolve } from "node:path";
 import { pathToFileURL } from "node:url";
 import { cosmiconfig } from "cosmiconfig";
@@ -26,7 +27,7 @@ const getDefaultSCSSAliases = (themePath: string): SCSSAliases => ({
 	],
 });
 
-export async function getSmashConfig(): Promise<Required<SmashConfig> | null> {
+export default async function getSmashConfig(): Promise<Required<SmashConfig> | null> {
 	const explorer = cosmiconfig("smash");
 	const config = await explorer
 		.load(resolve(process.cwd(), "smash.config.ts"))
@@ -103,7 +104,7 @@ export async function getSmashConfig(): Promise<Required<SmashConfig> | null> {
 						username: "",
 						password: "",
 					},
-				}
+				},
 			};
 			return defaultConfig;
 		});
@@ -120,7 +121,3 @@ function isValidSmashConfig(config: unknown): config is SmashConfig {
 		typeof config.themePath === "string"
 	);
 }
-
-
-// Fetch config vars from smash config
-
