+++ packages/smash-config/src/utils/getStagingUrl.ts
@@ -1,0 +1,10 @@
+import { SmashConfig } from "../";
+import { getConfigs } from "../";
+
+export default function getStagingUrl(config: SmashConfig): string {
+	const [value] = getConfigs(config, ["staging.url"]);
+	return value
+		.replace(/^https?:\/\//, "")
+		.replace(/^www\./, "")
+		.replace(/\/$/, "");
+}
